Fix path for `deno-proxy` tests
