+++ deno-proxy/README.md
@@ -6,18 +6,6 @@
 cargo install deno
 ```
 
-## Run JS Test
-
-```sh
-deno run --allow-env --allow-read ./deno-proxy/test.mjs
-```
-
-## Run Rust Test
-
-```sh
-cargo run --bin deno-proxy
-```
-
 ## Protocol
 
 The program communicates with `test.mjs` using STDIO. All messages should be ASCII symbols.
@@ -38,3 +26,17 @@ Each message contains
   ```
   8|{"a":42}11|[0,-1,true]
   ```
+
+## Debugging the `deno-proxy`
+
+### Run Rust `deno-proxy`
+
+```sh
+cargo run --bin deno-proxy
+```
+
+### Run JS `deno-proxy`
+
+```sh
+deno run --allow-env --allow-read ./deno-proxy/test.mjs
+```
